toxicity-detection-poc/services/decision-manager/requirements.txt
kafka-python==2.0.2
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.30.0
